Create monitor.py
The python script is used to measure the live power consumtion.

The consumed energy between two pulses is 3600 J. If the script measures the time T [s] between two pulses a good estimation of the live power consumption is   P = 3600 / T.

Usage:

python3 monitor.py

The script read data from GPIO pin 18 and sends the power consumption as a text message in a UDP package. The messages look like:
18 167.33
18 192.122
18 1200.8

These messages are then sent to the Blynk Repeater which is a Blynk client device and listens to UDP packages (which are pushed to the Blynk Server).
Usage :
Start Blynk Repeater with a device token (which is created in the mobile app). Then the device becomes online and ready to push data to the mobile app.
./BlynkRepeater -a 0d6219e705854b25bb535ccd489ee171

A message of the form
18 123.4

is pushed to virtual pin 18 (which is easily visualized in the mobile app).